+++ src/pages/request/index.jsx
@@ -5,6 +5,8 @@ import React from "react";
 function Request() {
     return (
         <div className="Request">
+                <img className="logo" src="/requestimg/logo.svg"/>
+                <img className="profile" src="/login/profile_female.svg"/>
             <Category firstCategory={'주문기기'} secondCategory={'핸드폰'}/>
             <Category firstCategory={'컴퓨터'} secondCategory={'기타'}/>
             <BottomBar />
