+++ src/styles/main.css
@@ -1,7 +1,7 @@
 .main1 {
     display: flex;
     justify-content: center;
-    margin-top: 286px;
+    margin-top: 200px;
     gap: 50px;
 }
 
@@ -84,9 +84,11 @@
     height: 55px;
     border-radius: 30px;
     color: white;
-    margin-top: 10px;
+    margin-left: 70px;
+    margin-top: 700px;
     border: none;
-    transition: all 1s;
+    z-index: 1;
+    position: absolute;
 }
 
 /* find */
@@ -95,7 +97,6 @@
     position: absolute;
     margin-top: 440px;
     margin-left: 125px;
-    z-index: 1;
 }
 
 .find-background {
@@ -120,7 +121,7 @@ nav > div {
 }
 
 .wrapper {
-    margin-top: 200px;
+    margin-top: 180px;
 }
 
 .img-wrapper {
@@ -136,9 +137,26 @@ nav > div {
 }
 
 .nav-word {
-    /* text-align: center; */
-    /* float: left; */
     width: 33%;
     height: 11px;
     line-height: 90px;
+}
+
+/*header*/
+
+.logo {
+    margin-left: 24px;
+    margin-top: 28px;
+    display: flex;
+    flex-direction: column;
+    top: 0;
+    left: 0;
+}
+
+.profile {
+    margin-top: -30px;
+    margin-left: 400px;
+    display: flex;
+    flex-direction: column;
+    top: 0;
 }
